nnscholar-search-main/requirements.txt
# Core dependencies
Flask>=2.0.0
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pandas==2.2.0
nltk==3.8.1
lxml==5.1.0
gunicorn==20.1.0

# AI Integration
openai>=1.12.0
anthropic>=0.42.0

# Data Visualization
matplotlib==3.8.2
seaborn==0.11.2
numpy==1.26.4

# Langchain
langchain-community==0.0.19
pydantic>=2.5.2
scikit-learn==1.4.0
openpyxl==3.1.2
